fix: support content page toggling (#4674)
* fix: support content page toggling

* fix: manage disclaimer md
