+++ sites/public/src/components/content-pages/PrivacySeeds.tsx
@@ -1,0 +1,31 @@
+import React, { useEffect, useContext } from "react"
+import { t } from "@bloom-housing/ui-components"
+import Markdown from "markdown-to-jsx"
+import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
+import { UserStatus } from "../../lib/constants"
+import Layout from "../../layouts/application"
+import pageContent from "../../md_content/privacy_policy.md"
+import { PageHeaderLayout } from "../../patterns/PageHeaderLayout"
+import styles from "../../patterns/PageHeaderLayout.module.scss"
+
+const Privacy = () => {
+  const { profile } = useContext(AuthContext)
+
+  useEffect(() => {
+    pushGtmEvent<PageView>({
+      event: "pageView",
+      pageTitle: "Privacy",
+      status: profile ? UserStatus.LoggedIn : UserStatus.NotLoggedIn,
+    })
+  }, [profile])
+
+  return (
+    <Layout>
+      <PageHeaderLayout heading={t("pageTitle.privacy")} inverse>
+        <Markdown className={styles["markdown"]}>{pageContent.toString()}</Markdown>
+      </PageHeaderLayout>
+    </Layout>
+  )
+}
+
+export default Privacy
